requirements.txt
PyPDF2
openai
python-dotenv
elevenlabs